Refactor: DTO 통합에 따른 변경사항 수정
