+++ src/main/java/sheetplus/checkings/domain/participatecontest/repository/ParticipateContestStateRepositoryCustom.java
@@ -1,7 +1,7 @@
 package sheetplus.checkings.domain.participatecontest.repository;
 
 import sheetplus.checkings.domain.member.dto.MemberDto.MemberInfoResponseDto;
-import sheetplus.checkings.domain.participatecontest.dto.response.ParticipateInfoResponseDto;
+import sheetplus.checkings.domain.participatecontest.dto.ParticipateContestDto.ParticipateInfoResponseDto;
 
 import java.util.List;
