+++ src/main/java/sheetplus/checkings/domain/participatecontest/repository/ParticipateContestStateRepositoryCustomImpl.java
@@ -4,7 +4,7 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import sheetplus.checkings.domain.member.dto.MemberDto.MemberInfoResponseDto;
-import sheetplus.checkings.domain.participatecontest.dto.response.ParticipateInfoResponseDto;
+import sheetplus.checkings.domain.participatecontest.dto.ParticipateContestDto.ParticipateInfoResponseDto;
 import sheetplus.checkings.domain.member.entity.Member;
 import sheetplus.checkings.domain.enums.MeritType;
